refactor: remove addContracts function and related routes; drop unused Team columns
